Добавит обработку офферов по клику на кнопку избранное
