+++ src/components/general-card/generalCard.tsx
@@ -1,7 +1,6 @@
 import {Link} from 'react-router-dom';
 import {Offer} from '../../types/offer';
 import {Card} from '../../types/card';
-//import {useAppDispatch, useAppSelector} from '../../hooks';
 import {FavoritesTriggerUpdate} from '../../const';
 import {useFavorites} from '../../hooks/useFavorites';
 
@@ -76,7 +75,7 @@ function GeneralCard({elementType, setActivePlaceCard, offer}: GeneralCardProps)
             <span className="place-card__price-text">&#47;&nbsp;night</span>
           </div>
           <button onClick={onChangeFavorites}
-            className={`place-card__bookmark-button ${offer.isFavorite ? 'place-card__bookmark-button--active' : ''} ${elementType === 'favorite' ? 'place-card__bookmark-button--active' : ''} button`}
+            className={`place-card__bookmark-button ${offer.isFavorite ? 'place-card__bookmark-button--active' : ''} button`}
             type="button"
           >
             <svg className="place-card__bookmark-icon" width="18" height="19">
