+++ src/pages/offer-page/offerPage.tsx
@@ -1,5 +1,5 @@
 import {useParams, Navigate} from 'react-router-dom';
-import {AppRoute, FavoritesTriggerUpdate} from '../../const';
+import {AppRoute} from '../../const';
 import {useState, useEffect} from 'react';
 import {useAppSelector} from '../../hooks/index';
 import Logo from '../../components/logo/logo';
@@ -14,7 +14,7 @@ import {getCity} from '../../store/offers-process/selectors';
 import {getOffer, getOfferIsLoading, getOfferIsNotFound} from '../../store/offer-process/selectors';
 import {getReviews} from '../../store/reviews-process/selectors';
 import {getOffersNearby, getOffersNearbyIsLoading} from '../../store/offers-nearby-process/selectors';
-import {useFavorites} from '../../hooks/useFavorites';
+import OfferNameWrapper from '../../components/offer/offer.tsx';
 
 const DEFAULT_BEGIN = 0;
 const MAX_IMAGES_SHOW = 6;
@@ -39,7 +39,6 @@ function OfferPage(): JSX.Element {
   const nearbyOffers = useAppSelector(getOffersNearby);
   const nearbyOffersIsLoading = useAppSelector(getOffersNearbyIsLoading);
 
-
   const [nearbyCardHoverId, setNearbyCardHoverId] = useState<string | null>(null);
 
   function handleCardHover(nearOfferId: string | null) {
@@ -51,13 +50,6 @@ function OfferPage(): JSX.Element {
   if(offerActive) {
     generalOffers.unshift(offerActive);
   }
-
-  const currentStatus = offerActive && offerActive.isFavorite ? 0 : 1;
-  const onChangeFavorites = useFavorites(
-    String(cardId),
-    currentStatus,
-    FavoritesTriggerUpdate.Offer
-  );
 
   return (
     <div className="page">
@@ -97,17 +89,7 @@ function OfferPage(): JSX.Element {
                     <span>Premium</span>
                   </div>
                   : ''}
-                <div className="offer__name-wrapper">
-                  <h1 className="offer__name">
-                    {offerActive.title}
-                  </h1>
-                  <button className="offer__bookmark-button button" type="button" onClick={onChangeFavorites}>
-                    <svg className={`offer__bookmark-icon ${offerActive.isFavorite ? 'offer__bookmark-button--active' : ''}`} width={31} height={33}>
-                      <use xlinkHref="#icon-bookmark" />
-                    </svg>
-                    <span className="visually-hidden">To bookmarks</span>
-                  </button>
-                </div>
+                <OfferNameWrapper cardId = {cardId} offerActive = {offerActive} />
                 <div className="offer__rating rating">
                   <div className="offer__stars rating__stars">
                     <span style={{ width: '80%' }} />
